Remove circular dep (#790)
